Work on better rewrite
As a proof of concept it worked well. This isn't going to replace ArchiSteamFarm or anything, but it does what it set out to do. I'll make it a lot simpler, and code it better with functions and all to make it more maintainable, then I'll leave it to become whatever it does.